Changes as specified in pull request
Changed to using native IBlockStates
Moved the config to a new class, TANConfig
Config is now validated against block states to check values specified
are valid.
Moved initialisation of TANConfig from preInit to init, as the config
now needs to have had all blocks for all mods created in preInit so that
it can validate the config against them.
The user now specifies an IBlockState to match against, and the set of
properties to consider for matching. If the properties list is empty
then only the block needs to match.
Changed case of FIRE to fire to be more consistent with project styling
Corrected indentation character from tabs to spaces to be more
consistent with project styling