no longer require use_properties
The config no longer has a field "use_properties" to specify which
properties of the state should be used to determine temperature.
Instead, the config can have properties in the state either set to a
wildcard ("*") or omitted. In both cases the effect is the same: any
value of that property is acceptable.