Fix Armors full stack (#862)
In the latest version, wool and leaves armor can be stacked in 64, I don't know if this is intentional but I suggest this change if it's not normal.
Normally, when you drag armors (Shift+Click), only one is equipped, but in a recent version, you just have to right-click to equip it automatically. 

With a stack of TAN armors, they're all equipped at once.